로그인 로직 옮김(Login.js -> app.js)
